CMFileManager: AOSP GET_CONTENT_DATA compatibility
This change brings compatibility to GET_CONTENT_DATA for AOSP apps when
using the PickerActivity:

* Detect crop extra; use the com.android.camera.action.CROP action of Gallery3d
to crop and return the requested image. This gets compatilibity for example with
the contacts app, when a user try to set the image of a contact.
* Detect android.provider.MediaStore.Audio.Media.EXTRA_MAX_BYTES; when this extra
is present the PickerActivity only display (and allow select) files with a size lower
than requested.
* Detect Intent.EXTRA_LOCAL_ONLY; useless until CMFM allow access remote file systems.

Change-Id: I1020458505b236653e869ec1c1f532dd6d686633