Issue 6795: File Manager selected file actions vanish on navigation to parent directory
Issue: http://code.google.com/p/cyanogenmod/issues/detail?id=6795

The issue is relate to back navigation (nor to ahead navigation). In this case the history
parcelable info data is recovery (which has a invalid selection data).
On back navigation, replace the selection data parcelable info with the current selection
of the navigation view.
Added notifyDataSetChanged() when toogle visibility, needed to refresh the view and repaint
the items.

Change-Id: Ib281c6fbc33274f6e81cc6b588cde420d6ceb19f